🐛 유효성 검사 추가 및 오버플로우 발생 버그 수정 bug (#162)
* :bug: 가격 계산시 Long형으로 변환하여 Overflow 방지

* :bug: Review와 Space에 유효성 검사 추가

+++ src/main/kotlin/com/beanspace/beanspace/domain/space/model/Space.kt
@@ -109,6 +109,6 @@ class Space(
     }
 
     fun calculateTotalCost(headCount: Int, stayDays: Long): Long {
-        return (this.price + (headCount - this.defaultPeople).coerceAtLeast(0) * this.pricePerPerson) * stayDays
+        return (this.price + (headCount - this.defaultPeople).coerceAtLeast(0) * this.pricePerPerson.toLong()) * stayDays
     }
 }
